backend/requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-multipart

# LangChain and AI/ML
langchain
langchain-chroma
langchain-core
langchain-ollama
langchain-community
langchain-openai
langchain-text-splitters
chromadb

# Document processing
unstructured[all-docs]
pypdf
python-pptx
python-docx
openpyxl

# Web scraping
beautifulsoup4
requests
httpx

# Utilities
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dateutil
PyYAML

# Development
debugpy
pytest
pytest-asyncio
httpx

# Optional (for production)
gunicorn
uvloop
websockets
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
mdurl
mmh3
mpmath
narwhals
numpy
oauthlib
ollama
onnxruntime
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
pillow
posthog
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
pytz
PyYAML
referencing
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
streamlit
sympy
tenacity
tokenizers
toml
tornado
tqdm
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchdog
watchfiles
websocket-client
websockets
wheel
wrapt
zipp
zstandard
